+++ CHANGELOG.md
@@ -15,6 +15,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Security
 -->
 
+
+
+## v1.10.1
+### Changed
+- nc-slideshow buttons now appear or disappear depending on the number of slide displayed
+
+## v1.10.0
+### Added
+- nc-slideshow button next/previous now have slot to add any icon
+
+
 ## v1.9.3
 ### Added
 - nc-text-input error tag with parameters